# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the do-it swagger docs",
        "title": "Do-It Swagger",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Do-It Support",
            "url": "https://github.com/milwad-dev",
            "email": "milwad.dev@gmail.com"
        },
        "version": "1.0"
    },
    "basePath": "/api/",
    "paths": {
...
        "/api/labels": {
            "get": {
                "description": "Get the latest labels",
                "produces": [
                    "application/json"
                ],
                "summary": "Get Labels",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Label"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "store new label",
                "produces": [
                    "application/json"
                ],
                "summary": "Store Label",
                "parameters": [
                    {
                        "description": "The title of the label",
                        "name": "title",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "The color of the label",
                        "name": "color",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/labels/{id}": {
            "delete": {
                "description": "delete label by id",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete Label",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "The id of the label",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Label": {
            "type": "object",
            "required": [
                "color",
                "title"
            ],
            "properties": {
                "color": {
                    "type": "string",
                    "maxLength": 250,
                    "minLength": 3
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string",
                    "maxLength": 250,
                    "minLength": 3
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "required": [
                "email",
                "id",
                "name",
                "phone"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string",
                    "maxLength": 250,
                    "minLength": 3
                },
                "emailVerified_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "maxLength": 250,
                    "minLength": 3
                },
                "phone": {
                    "type": "string"
                },
                "phoneVerified_at": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}